created separate go.mod file for listener module
